pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pinelabs-mcp-server"
version = "0.1.0"
description = "Official Pine Labs MCP Server for Pine Labs payment APIs."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Pine Labs" }
]
keywords = ["mcp", "pinelabs", "payments", "fastmcp"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "dateparser>=1.2.0",
  "fastmcp>=3.2.0",
  "httpx>=0.27.0",
  "mcp>=1.0.0",
  "pydantic>=2.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
]

[project.scripts]
pinelabs-mcp-server = "cli.pinelabs_mcp_server.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["cli*", "pkg*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]